relationsship between the tables

+++ src/main/java/dk/kea/kinoxp_rest/dto/SeatDTO.java
@@ -2,6 +2,6 @@ package dk.kea.kinoxp_rest.dto;
 
 import dk.kea.kinoxp_rest.model.Theater;
 
-public record SeatDTO(int seat_ID, String seat_number, boolean isReserved, Theater theater, int seat_Price) {
+public record SeatDTO(int seat_ID, String seat_number, boolean isReserved, Theater theater, double seat_Price) {
 
 }
