+++ src/main/java/dk/kea/kinoxp_rest/model/Seat.java
@@ -15,7 +15,7 @@ public class Seat {
     private int seat_ID;
     private String seat_number;
     private boolean isReserved;
-    private int seat_Price;
+    private double seat_Price;
     //private int theater_ID;
 
 
@@ -25,5 +25,8 @@ public class Seat {
     @ManyToOne
     @JoinColumn(name = "theater_idfk")
     private Theater theater;
+
+
+
 
 }
